change: check if place contains any item before deleting it

+++ src/main/java/cloud/ptl/indexer/api/place/PlaceService.java
@@ -36,6 +36,10 @@ public class PlaceService {
     }
 
     public void deletePlace(Long id) {
+        PlaceEntity place = getPlace(id);
+        if (!place.getItems().isEmpty()) {
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, "Place is not empty");
+        }
         placeRepository.deleteById(id);
     }
